make nudge, end inherit one class because they dont differ in mtef3/5
