+++ lib/mathtype.rb
@@ -2,7 +2,7 @@ require "mathtype/version"
 require "bindata"
 require "ole/storage"
 require "nokogiri"
-require_relative "records/bintypes.rb"
+require_relative "records/mtef.rb"
 require_relative "records3/mtef.rb"
 require_relative "records5/mtef.rb"
