+++ lib/records/snapshot.rb
@@ -1,0 +1,19 @@
+require "bindata"
+
+module Mathtype
+  module Snapshot
+    def snapshot
+      snapshot = BinData::Record::Snapshot.new
+      exposed = self.class.const_get(:EXPOSED_IN_SNAPSHOT)
+      exposed.each do |name|
+        obj = find_obj_for_name(name)
+        if obj
+          snapshot[name] = obj.snapshot if include_obj?(obj)
+        else
+          snapshot[name] = self.send(name)
+        end
+      end
+      snapshot
+    end
+  end
+end
