+++ lib/records3/line.rb
@@ -22,7 +22,7 @@ module Mathtype3
 
     virtual :_tag_options, :value => lambda{ _options }
 
-    nudge :nudge, onlyif: lambda { _options & OPTIONS["xfLMOVE"] > 0 }
+    record_nudge :nudge, onlyif: lambda { _options & OPTIONS["xfLMOVE"] > 0 }
 
     mtef16 :line_spacing, onlyif: (lambda do
       _options & OPTIONS["xfLSPACE"] > 0
