+++ lib/records3/tmpl.rb
@@ -721,7 +721,7 @@ module Mathtype3
 
     virtual :options, :value => lambda{ _options }
 
-    nudge :nudge, onlyif: lambda { _options & OPTIONS["xfLMOVE"] > 0 }
+    record_nudge :nudge, onlyif: lambda { _options & OPTIONS["xfLMOVE"] > 0 }
 
     int8 :_selector
